refactor: extract ShuffleService (has move / shuffle)
