+++ Assets/Scripts/Board/Services/ShuffleService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19a0ec31d53ca48f2a1cd121b6b119f0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
